Correct VMwrtbarWithoutStoreEvaluator's handling of ops besides wrtbari
In dealing with improvements to wrtbari, I had messed up how VMwrtbarWithoutStoreEvaluator was generating code for other opcodes.  For wrtbari, I had pulled a "tenured" check of the target to be the outermost condition generated, and replaced it with a source value check or no check in a potentially nested position.  But that resulted in no tenure check and an unconditional call to the VM helper for many opcodes.  Reworked so that potentially nested check of whether something is in tenure can deal with either the source value (for wrtbari) or the target (for other opcodes).

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>

WIP: #2135