Look for additional opportunities for stack allocation
This introduces some further refinements to detect opportunities for
Escape Analysis (EA) in loops where references are copied between local
variables.

Firstly, this change considers the right-hand side of an assignment to
be harmless for the purposes of EA if it was an object that was locally
allocated by a previous pass of EA, not only if it is a candidate for
local allocation in the current pass of EA.

Next, any uses of an object that occur under a call to prepareForOSR
under voluntary OSR can be ignored when checking for the uses reached
by each definition.

These changes were prototyped by Andrew Craik <ajcraik@ca.ibm.com>.

Also, fixed a bug in trace code that was getting the class name of a
TR::New operation incorrectly.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>