Use node global index to access entries in _ignoreableUses
The code in findIgnoreableUses that populates the TR_BitVector,
_ignoreableUses, uses the global indices of nodes to index the
TR_BitVector.  Code in checkDefsAndUses was using the value numbers of
nodes to check the entries in _ignoreableUses.

Fixed checkDefsAndUses to use nodes' global indices instead.  Also,
reorganized the code checking _ignoreableUses to pull the invariant
check out of a loop.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>