Look for right caller method when handling OSR in context of inliner
In many places, OSR-related code in the inliner is calling
getOwningMethodSymbol() on a symbol reference in order to access
information about the calling method.  However, the owning method for
a symbol reference is not necessarily the calling method.

Instead, code should call getInlinedResolvedMethodSymbol() on the
TR::Compilation object for the bytecode's caller index or, if the caller
index is -1, call getMethodSymbol().

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>