Define separate enum constants for recognized System.arraycopy methods
The handling of System.arraycopy as a recognized method uses a wildcard
for its signature.  We will need to distinguish the two methods, so this
change introduces distinctly named enum constants for them.  These will
have the same value for now, as OMR has references to the single old
constant name.

Once OMR is updated to refer to the new names, the
defintion of the old enum constant can be dropped, and the values of the
two new constant names made distinct.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>