Check whether recognized methods expected to be native actually are
Enable checking of whether a recognized method that is expected to
be native actually is native.  If it is not, the compiler should not
recognize it.

Setting the TR_CheckRecognizedMethodNativeness environment variable
enables a fatal assertion check that a recognized method is expected
to be native if and only if it actually is native.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>