Define new macros for initialization of recognized method entries
Add native version fields to definition of struct X for recognized
method table entries.  These define a range of Java versions over
which a recognized method is expected to be native.

The entries can be initialized using one of a set of macros that supply
default values for recognized methods that are not expected to be
native or to actually supply values for a range of Java versions.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>