Avoid functions that might not be folded in static initializers
When compiling with xlC on z/OS or Power, references to
std::numeric_limits<>::{min|max}() might not be folded at comple-time.
As references to macros that contain these functions are being used
quite extensively, we need to avoid the impact of evaluating all those
function references at start-up time.  Replacing those with references
to INT8_MIN and INT8_MAX.

Also replacing references to strlen in static initializers with calls
to sizeof()-1 for the same reason.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>