Optimize object equality comparison in VP
Eqaulity comparisons can be optimizing by creating and using VP
constraints.

First, the objectEqualityComparison non-helper can be constrained to
have a value of either 0 or 1, by definition.

Second, the non-helper call can be decayed to a regular acmpeq node
under the following cases:

- when either operand is known to be a non-value type, a reference
  (or identity) comparison is performed (even when the other operand
  is a value type because the comparison always evaluates to false)
- when the two operands are constrained to have the same reference

Further constraints are handled recursively on the acmpeq node, if
transformed.

Signed-off-by: Leonardo Banderali <leonardo2718@protonmail.com>