Ensure that jitNewObject for value type is not evaluated inline
If the class passed to the jitNewObject helper is known to be a value
type, have canAllocateInlineClass returns false for it.  That prevents
the JIT from generating inline code for allocating an instance,
including preventing it from being eligible for escape analysis.  That
ensures the helper is called at runtime, allowing it to throw an
InstantiationError.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>