Avoid inline allocation of arrays of value types
An array whose elements are of a value type needs to be initialized
with the default value of its elements' value type when it is
created with an ANEWARRAY or MULTIANEWARRAY bytecode.  Inline
allocation of arrays is not yet able to handle this, so inline
allocation has to be prevented for such arrays.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>