Fix init sequence for flattened fields
In cases where a flattened field is contained in an identitytype, there
is a bug where the flattenedfield is not being initialized before the
container type is initialized. This is because there is an unneccesary
valuetype guard before that code. Since identitytypes can have flattened
fields, identity type must also ensure that all flattened fields are
initialized before itself.

The flattenedclass cache will be non-NULL if the type is a valuetype or
the type has at least one flattened field. This means that one simply
needs to check if the flattenedClass cache is non-NULL before
initializing flattened fields.

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>