Fix minor problems for VP support of object equality helper
Test of whether an object could hold an instance of a value type should
consider whether its type is not concrete - i.e., not abstract and not
an interface.  Also, constraints for object equality comparison should
only be added during last time through, in case the call is in a loop.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>