Add inline check to aaload and aastore in treeLowering
Currently when value type is enabled, ILGen generates helper calls
jitLoadFlattenableArrayElement and jitStoreFlattenableArrayElement
for both VT and non-VT aaload and aastore. This approach causes
performance regression. This change introduces inline checks in
tree lowering.
Non-VT aaload and aastore blocks are added as an extended block of
the original block that contains the helper call. The helper call is
split into its own block and moved to the end of the tree list.
Runtime check on the array component class is inserted before aaload
and aastore operations. If it's value type, the control flow jumps
to the out of line helper call block. If not, the control falls
through to the non-VT aaload and aastore blocks.

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>