Check for store of non-value type to array element
If a value that is to be stored to an array element via a call to the
jitStoreFlattenableArrayElement helper is not a value type, value
propagation can transform it to use inline code.  If the array has a
value type as its component type, an ArrayStoreCheckException or a
NullPointerException will result, so there is no need to use the helper
to handle the possibility of a flattenable array.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>