Define symbol reference for value type array null store check
Define a new method to find or create a symbol reference for a non-
helper method that will be used to check whether an array is a
value-types array, and if so, whether a null value is being stored
to one of its elements.  If so, a NullPointerException will be
thrown.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>