Add NULL check before checking if array component type is VT
When generating inline IL for nonNullableArrayNullStoreCheck in the Tree
Lowering optimization, rather than checking whether the component type
of an array is a value type and then checking whether the value being
stored is null, it should typically be faster to check first whether the
value is null.  That avoids the double indirection to access the class
flags of the component type if the value that's being assigned is non-
null.

Also added missing edges to CFG for branches around NULLCHK.

Authored-by: Annabelle Huo <Annabelle.Huo@ibm.com>

Signed-off-by: Henry Zongaro <zongaro@ca.ibm.com>