Allow inlining java/lang/ref/Reference.refersTo method
A Reference.refersTo(Object) method, introduced in Java 16, tests
whether the referent is the same object as the parameter.  Some
places in the JCL that used to perform comparisons of the referent using
Reference.get now use Reference.refersTo.  That affects performance of
such uses as Reference.get was implemented using a call to
Reference.getImpl which the JIT compiler recognized and inlined if
possible.

This change adds recognition of Reference.refersTo to the JIT, which
will now inline a call to it under the same conditions as it would
inline a call to Reference.getImpl.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>