Define queries to test whether Reference.refersTo can be inlined
The refersTo method was introduced to java.lang.ref.Reference in
Java 16.  The JIT compiler is able to inline the method under the same
conditions as it is able to inline Reference.getImpl.  This change
introduces new query methods and fields to check and record whether both
methods can be inlined.

The existing VM::isGetImplInliningSupported and
Compilation::setGetImplInlineable methods are invoked by OMR, so
they will remain in OpenJ9 until a subsequent pull request against OMR
replaces calls to those now obsolete methods.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>