Compare classes in tree lowering reference comparison fast path
In Tree Lowering for <object{E|Ine}qualityComparison>, the last two fast path checks test whether either operand of the comparison is a value type, and if not, sets the result to unequal.  However a necessary condition for two value types to be equal is for them to have the same class, so a slightly cheaper test is to check whether the first operand is a value type and whether the two, blah, blah, blah.