Check for harmless calls in checking for harmless stores
In the process of checking a candidate for stack allocation in a loop,
escape analysis examines locals into which a reference to the candidate
might have been copied and all the other definitions that might reach
that point to see whether they are considered harmless.  This change
adds method calls to the list of things considered harmless if the
arguments to the method are not addresses (other than the result of
aconst or loadaddr operation), as such a method call cannot result in
a reference to the candidate for stack allocation from a previous
iteration being returned.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>