Make minor clean up changes to Escape Analysis
This change implements several minor clean-up items for Escape Analysis:

1) The top-level findIgnoreableUses method looks for calls to
eaEscapeHelper and then calls a second method named findIgnoreableUses
to mark uses of candidates for stack allocation as ignoreable.  However,
that second method doesn't actually find ignoreable uses, it just marks
them.  This change renames that second method to markUsesAsIgnoreable
for clarity.

2) Most trace code in Escape Analysis identifies a candidate for stack
allocation by the address of the TR::Node for the allocation, which is
stored in the _node field of a Candidate object.  In one case, tracing
was printing the address of the Candidate object itself.

3) Removed check of environment variable that is used to disable
checking in loops for aliases of candidates for stack allocation, and
removed method checkAllNewsOnRHSInLoop that was used in that case.  The
code that it allows to be disabled has been in place for about three
years now, so it seems relatively safe.

4) Added dynamic debug counters to guage number of contiguous and
non-contiguous stack allocation opportunities that were found.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>