Move code that checks for definitions in a loop
The checkAllNewsOnRHSInLoopWithAliasing method looks at locals that
could refer to a candidate for stack allocation that's allocated in a
loop in the IL, and it checks a number of conditions on all the
definitions that reach that use tosee verify that they are considered
harmless.  This change moves those checks into a new method,
isHarmlessDefInLoop, in order to make subsequent changes easier to
understand.

The change should have no functional or performance difference.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>