Create unit tests for potential division by zero exceptions
This change defines some unit tests for int and long division and
remainder operations that could result in ArithmeticExceptions for a
division by zero.  These tests are intended to exercise various
situations in which optimization of division and remainder operations is
known to occur to ensure checks for division by zero remain if needed.
It also tests some complicating scenarios, such as commoning of operands
or operands of division or remainder operations that are themselves the
results of division or remainder operations.