Check for null parent in TR_EscapeAnalysis::fixupNode
In TR_EscapeAnalysis::fixupNode, code that rewrites an awrtbari as an
astorei assumes that the awrtbari node must have a parent node, but
that's not necessarily the case.  Added a test of whether the parent
is NULL prior to checking whether the parent is a check operation or
a compressedrefs.

Further, in recreating the parent as a TR::treetop, fixupNode needs to
take into account the possibility that the parent has children other
than the awrtbari that are no longer needed once the parent is recreated
as a TR::treetop.