Use signed comparisons for Latin-1 encoded indexOf
The implementations of String.indexOf operations cast the target char to
a byte if Latin-1 encoded Strings are involved.  The value might be cast
to type byte by the caller or in the implementation itself.  Value
Propagation has to ensure that it uses signed byte comparisons if it
transforms an indexOf involving a known String.