Inline java.lang.Class.getComponentType() in Value Propagation
Check for the opportunity to inline calls to Class.getComponentType,
using the underlying j9ArrayClass.  In particular, if the Class is known
to be an array class, the test for the J9AccClassArray flag in the
classDepthAndFlags field can be skipped.

Also, mark Object.getClass() as non-null during IL Generation.  That
prevents a NULLCHK from being being generated for the Class reference in
an expression like o.getClass().getComponentType().  The presence of the
NULLCHK would inhibit this transformation.