REVISED to co-exist with static field handling Change order of tests for inline Unsafe.{get|put}* calls
If the J9Class for java/lang/Class is available to the compiler, the
IL it generates inline for Unsafe.{get|put}{int|long|Object}* tests
whether the Object being accessed is null (for native memory
references), has the low-order bit set for the specified offset
and whether its J9Class is that of java/lang/Class, in that order.
Those tests are trying to distinguish whether a static field is being
accessed, which requires different code to access the memory than
all the other cases.

For both instance fields and static fields, the Object will be a non-
null reference.  However, the low-order bit of the offset will never
be set for instance fields.  Therefore, what is likely the most common
usage would benefit from having the low-order offset bit tested first.

This commit changes the order of the tests accordingly for that case.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>