Adjust expected return type for calls to getStringUTF8Length
The return type of the JVM's getStringUTF8Length method has changed from
IDATA to UDATA.  This change adjusts the JIT compiler's uses of that
method.  In particular, the return type of
TR_FrontEnd::getStringUTF8Length and its overriding methods changes from
intptr_t to int32_t.  Similarly, the bufferSize argument of
TR_FrontEnd::getStringUTF8 becomes uintptr_t where it was int32_t.

The motivation was that the length of a UTF-8 encoded String could be
greater than 2^32 bytes, so the length could overflow on a 32-bit
platform.  All uses of getStringUTF8Length in the JIT all involve
signatures, descriptors, method names and class names, which will never
be large enough to exceed the range of the int32_t type.  Just to be
cautious, however, this change includes an assertion test that the
computed length is in range for the int32_t type.

This change also introduces a method, getStringUTF8UnabbreviatedLength,
that returns a length value of type uint64_t if the JIT compiler ever
needs to determine the UTF-8 encoded length of an arbitrary String.  The
method is currently unused.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>