Throw IdentityException if operand of monent is a value type object
Previously, specifying an instance of a value type class on a
monitorenter bytecode would result in an IllegalMonitorStateException;
a java.lang.IdentityException must now be thrown instead.

This commit introduces a non-helper, <isIdentityObject>, that checks
that the operand of a TR::monent instruction is an instance of an
identity class rather than a value type class.  If it is a value type
class, it throws an IdentityException.  The actual test is expanded
to test the flags field of the object's class during the Tree Lowering
optimization pass.

The result of <isIdentityObject> is tested on a ZEROCHK operation that
is generated before the monent during IL generation.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>