Address review comments
The newly introduced Compilation Server message type
VM_getStringUTF8UnabbreviatedLength is never acutally used.  This change
removes it along with the VM_getStringUTF8Length message type, which is
also never used.

This change also adjusts the maximum length for a UTF-8 encoding that
can be returned by getStringUTF8Length from 2^31-1 to 2^31-2.  That
ensures that any code that then uses that length to allocate buffer to
contain the encoded String with a NUL terminator will not overflow a
32-bit signed integer representation for the length plus the NUL byte.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>