Remove ifdefs from uses of jitThrowIdentityException
Two references to jitThrowIdentityException were conditionally compiled
under a test of whether value types support is enabled.  That is not
necessary as the helper functions jitThrowIdentityException and
old_slow_jitThrowIdentityException are defined regardless of whether
value types support is enabled.

Also corrected the spacing of an operation in an example of the
transformation of the IL for calls to the <isIdentityObject> non-helper.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>