Fix non-contiguous field only if base value number matches candidate
For a reference to a field whose value number is one that might be used
by a non-contiguous candidate, code in TR_EscapeAnalysis::fixupNode
called fixupFieldAccessForNonContiguousAllocation unconditionally.  That
would always transform the indirect field reference into a direct
reference to the automatic variable representing that field of the
candidate.

However, a candidate's value numbers can include references that might
access the candidate, but do not necessarily.  For most objects, that
would prevent the candidate from being stack allocated non-contiguously,
but immutable objects can still be allocated discontigously even in the
presence of references that could be reached by objects.

This change will only call fixupFieldAccessForNonContiguousAllocation if
the value number of the non-contiguous candidate's allocation node
matches the value number of the base of this indirect reference.  That
ensures the reference is definitely a reference to a field of that
candidate.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>