Detect any monitorenter before point of dynamic loop transfers
Dynamic loop transfer can transfer control from the interpreter into a
JIT-compiled method after the point at which monitorenter bytecode
instructions have been executed.  With Java 24 and later, the JVM uses
the JIT's live monitor metadata to locate monitors that might need to be
detached from carrier threads when the owning virtual thread is
unmounted.

If the JIT is unable to determine that no monitorenters will be executed
prior to the entry point for dynamic loop transfer, it must fail the
compilation.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>