Check JVM flag for isLiveMonitorMapCorrectnessRequired
The JVM only requires live monitor metadata to be correct if the
J9_EXTENDED_RUNTIME3_YIELD_PINNED_CONTINUATION flag is set in
extendedRuntimeFlags3.  This change adjusts the implementation of
TR_J9VMBase::isLiveMonitorMapCorrectnessRequired to return true only if
that flag is set.  It can be turned off by specifying the option
-XX:-YieldPinnedVirtualThreads.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>